fixed some errors of the program

+++ sprint3/src/onlineStore/pendingStoreHandler.java
@@ -19,27 +19,29 @@ public class pendingStoreHandler {
 	public void approveStore(String storeName) throws ClassNotFoundException, SQLException {
 		String q1="select * from tempStore where name = '"+storeName+"'";
 		db.rs = db.st.executeQuery(q1);
-	    db.rs.next();
-	    int ID = db.rs.getInt("ID");
-	    String name = db.rs.getString("name");
-	    String username = db.rs.getString("ownerUser");
-	    String location = db.rs.getString("location");
-	    String type = db.rs.getString("type");
-	    System.out.println("got here");
-		String q2="select ID from users where username ='"+username+"' ";
-		db.rs = db.st.executeQuery(q2);
-		db.rs.next();
-		int ownerID = db.rs.getInt(1);
-		String q3 = "insert into store values(" +ownerID + ", '" +location+"', '" +type+"', '" +name+ "')";
-		db.st.executeUpdate(q3);
-		String q4 = "select store.ID from store where name = '"+name+ "' and owner = "+ownerID;
-		db.rs = db.st.executeQuery(q4);
-		db.rs.next();
-		int storeID = db.rs.getInt(1);
-		String q5 = "insert into collaborators values("+ownerID+","+storeID+")";
-		db.st.executeUpdate(q5);
-		String q6 = "delete from tempStore where ID = "+ID;
-		db.st.executeUpdate(q6);
+	    if(db.rs.next())
+	    {
+	    	int ID = db.rs.getInt("ID");
+	    	String name = db.rs.getString("name");
+	    	String username = db.rs.getString("ownerUser");
+	    	String location = db.rs.getString("location");
+	    	String type = db.rs.getString("type");
+	    	System.out.println("got here");
+	    	String q2="select ID from users where username ='"+username+"' ";
+	    	db.rs = db.st.executeQuery(q2);
+	    	db.rs.next();
+	    	int ownerID = db.rs.getInt(1);
+	    	String q3 = "insert into store values(" +ownerID + ", '" +location+"', '" +type+"', '" +name+ "')";
+	    	db.st.executeUpdate(q3);
+	    	String q4 = "select store.ID from store where name = '"+name+ "' and owner = "+ownerID;
+	    	db.rs = db.st.executeQuery(q4);
+	    	db.rs.next();
+	    	int storeID = db.rs.getInt(1);
+	    	String q5 = "insert into collaborators values("+ownerID+","+storeID+")";
+	    	db.st.executeUpdate(q5);
+	    	String q6 = "delete from tempStore where ID = "+ID;
+	    	db.st.executeUpdate(q6);
+	    }
 	}
 	public ArrayList<store> viewPendingApproves(){
 		String q1="select * from tempStore";
